test: organize storage conformance tests
